+++ src/indicators/__init__.py
@@ -2,6 +2,18 @@
 
 from indicators.fvg import FvgConfig, FvgMitigation, FvgSignal, detect_fvgs
 from indicators.ifvg import IfvgConfig, IfvgPoi, IfvgSignal, detect_ifvgs
+from indicators.structure import (
+    BosEvent,
+    DealingRange,
+    StructureConfig,
+    StructureResult,
+    StructureState,
+    SwingPoint,
+    analyze_structure,
+    detect_bos,
+    detect_swings,
+    derive_structure_state,
+)
 
 __all__ = [
     "FvgConfig",
@@ -10,6 +22,16 @@ __all__ = [
     "IfvgConfig",
     "IfvgPoi",
     "IfvgSignal",
+    "BosEvent",
+    "DealingRange",
+    "StructureConfig",
+    "StructureResult",
+    "StructureState",
+    "SwingPoint",
+    "analyze_structure",
+    "detect_bos",
+    "detect_swings",
     "detect_fvgs",
     "detect_ifvgs",
+    "derive_structure_state",
 ]